First draft of completed search algorithm. An exhaustive search of all possible ways the human could play passes completely: there is no way to win.
Some parts need to be cleaned up. The heuristics at the top of my `computer_play()` method do save a lot of unnecessary recursive calls to the minimax tree, but a fair amount of calls to it are still occurring. An easy way to speed it up would be memoization, but considering it can play every possible game in 0.33s, I'm not terribly worried about it at this point.